move files to root , change clases name, etc
