+++ src/Api/Operations/Request.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Canvas\Api\ApiOperations;
+namespace Canvas\Api\Operations;
 
-use Canvas\Api\ApiRequestor;
+use Canvas\Api\Requestor;
 use Canvas\Util\RequestOptions;
 
 /**
@@ -59,7 +59,7 @@ trait Request
     {
         $requestOptions = RequestOptions::parse($options);
         $baseUrl = isset($requestOptions->apiBase) ? $requestOptions->requestOptions : static::baseUrl();
-        $requestor = new ApiRequestor($requestOptions->apiKey, $baseUrl);
+        $requestor = new Requestor($requestOptions->apiKey, $baseUrl);
         list($response, $requestOptions->apiKey) = $requestor->request($method, $url, $params, $requestOptions->headers, $requestOptions->query);
         $requestOptions->discardNonPersistentHeaders();
         return [$response, $requestOptions];
