--- src/Resources/Users.php
@@ -1,23 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Canvas\Resources;
-
-use Canvas\Api\ApiOperations\All;
-use Canvas\Api\ApiOperations\Create;
-use Canvas\Api\ApiOperations\Delete;
-use Canvas\Api\ApiOperations\Update;
-use Canvas\Api\ApiOperations\Retrieve;
-use Canvas\Api\ApiResource;
-
-class Users extends ApiResource
-{
-    const OBJECT_NAME = 'users';
-
-    use All;
-    use Create;
-    use Delete;
-    use Update;
-    use Retrieve;
-}
